+++ htdocs/timesheet/class/TimesheetUserTasks.class.php
@@ -568,11 +568,13 @@ public function updateActuals($tabPost, $notes = array())
         /*
          * For each task store in matching the session timestamp
          */
-        foreach($this->taskTimesheet as $key  => $row) {
-            $tasktime = new TimesheetTask($this->db);
-            $tasktime->unserialize($row);
-            $ret+=$tasktime->postTaskTimeActual($tabPost[$tasktime->id], $this->userId, $this->user, $this->timestamp,  $notes[$tasktime->appId]);
-            $this->taskTimesheet[$key] = $tasktime->serialize();
+        if(is_array($this->taskTimesheet)){
+            foreach($this->taskTimesheet as $key  => $row) {
+                $tasktime = new TimesheetTask($this->db);
+                $tasktime->unserialize($row);
+                $ret+=$tasktime->postTaskTimeActual($tabPost[$tasktime->id], $this->userId, $this->user, $this->timestamp,  $notes[$tasktime->id]);
+                $this->taskTimesheet[$key] = $tasktime->serialize();
+            }
         }
         /*
     if(!empty($idList)) {
@@ -765,14 +767,14 @@ public function getHTMLHeader()
         if(count($this->headers) == 1) {
                 $html .= 'colspan = "2" ';
         }
-        $html .= ">".$langs->trans($value)."</th>\n";
+        $html .= "> <a onclick=\"sortTable('timesheetTable_{$this->id}','col{$value}','asc');\">".$langs->trans($value)."</a></th>\n";
     }
     $opendays = str_split($conf->global->TIMESHEET_OPEN_DAYS);
     for ($i = 0;$i<$weeklength;$i++)
     {
         $curDay = $this->date_start+ SECINDAY*$i+SECINDAY/4;
         $htmlDay = ($conf->global->TIMESHEET_TIME_SPAN == "month")?substr($langs->trans(date('l', $curDay)), 0, 3):$langs->trans(date('l', $curDay));
-        $html .= "\t".'<th class = "days_'.$this->id.'" id = "'.$this->id.'_'.$i.'" width = "35px" style = "text-align:center;" >'.$htmlDay.'<br>'.dol_print_date($curDay, $format)."</th>\n";
+        $html .= "\t".'<th class = "daysClass days_'.$this->id.'" id = "'.$this->id.'_'.$i.'" width = "35px" style = "text-align:center;" >'.$htmlDay.'<br>'.dol_print_date($curDay, $format)."</th>\n";
     }
     return $html;
 }