+++ htdocs/timesheet/core/js/timesheet.js
@@ -154,13 +154,13 @@ function removeUnchanged(){
         
         var lineClass="line_"+tsUser[j].value;
         //foreach task
-        var task= document.getElementsByClassName(lineClass);
-        var nbTask=task.length;
+        var task = document.getElementsByClassName(lineClass);
+        var nbTask = task.length;
         for(i=0;i<nbTask;i++){  
             changed=0;
-            var inputs=task[i].getElementsByTagName( 'input' );
-            var textarea=task[i].getElementsByTagName( 'textarea' );
-            var nbInputs=inputs.length;
+            var inputs = task[i].getElementsByTagName( 'input' );
+            var textarea = task[i].getElementsByTagName( 'textarea' );
+            var nbInputs = inputs.length;
             var nbTextarea=textarea.length;
             for(k=0;k<nbInputs;k++){
                 if(inputs[k].defaultValue!=inputs[k].value)changed++
@@ -382,7 +382,7 @@ function searchTask(evt){
         var displayLine=(tslist[i].id=="searchline")?true:false;
         for (j=0; j<fields.length;j++){
            var found=0;
-           found+=fields[j].innerText.search(search);
+           found+=fields[j].innerHTML.search(search);
            if(found>=0){
              displayLine=true;
             }
@@ -558,8 +558,13 @@ function openNote(noteid){
 //function to close note
 function closeNotes(){
     var modals = document.getElementsByClassName("modal");
-    for(var modal in modals){
-        (modals[modal]).style.display = "none";
+    var patt = /(\w+)\.png$/gi 
+    for(var i=0;i<modals.length;i+=1){
+        var modalbox = modals[i];
+        modalbox.style.display = "none";
+        var icon = (modalbox.firstChild.lastChild.value.length>0)?"file":"filenew";
+        var imgnote = document.getElementById("img_"+modalbox.id);
+        imgnote.src = imgnote.src.replace(patt,"$'"+icon+".png");
     };
 }
 
@@ -571,3 +576,45 @@ window.onclick = function(event) {
         //closeNotes();
     }
 }
+
+// https://www.w3schools.com/howto/howto_js_sort_table.asp
+function sortTable(table,col,sort) {
+    var table, rows, switching, i, x, y, shouldSwitch;
+    table = document.getElementById(table);
+    switching = true;
+    /* Make a loop that will continue until
+    no switching has been done: */
+    while (switching) {
+      // Start by saying: no switching is done:
+      switching = false;
+      rows = table.getElementsByClassName('timesheet_line');
+      /* Loop through all table rows (except the
+      first, which contains table headers): */
+      for (i = 1; i < (rows.length - 1); i++) {
+        // Start by saying there should be no switching:
+        shouldSwitch = false;
+        /* Get the two elements you want to compare,
+        one from current row and one from the next: */
+        x = rows[i].getElementsByClassName(col)[0];
+        y = rows[i + 1].getElementsByClassName(col)[0];
+        // Check if the two rows should switch place:
+        if(typeof(x) !== 'undefined' && typeof(y) !== 'undefined'){
+            if (sort == "desc" && (x.innerHTML.toLowerCase() > y.innerHTML.toLowerCase())){
+                // If so, mark as a switch and break the loop:
+                shouldSwitch = true;
+                break;
+            }else if(sort == "asc" && (x.innerHTML.toLowerCase() < y.innerHTML.toLowerCase())){
+                // If so, mark as a switch and break the loop:
+                shouldSwitch = true;
+                break;                
+            }
+        }
+      }
+      if (shouldSwitch) {
+        /* If a switch has been marked, make the switch
+        and mark that a switch has been done: */
+        rows[i].parentNode.insertBefore(rows[i + 1], rows[i]);
+        switching = true;
+      }
+    }
+  }